Streamline database tools to single nightkidz-db-setup.js script and update documentation
