--- SOLUTION.md
@@ -1,65 +1,0 @@
-# Solution: Local Development Tools Without Affecting Deployment
-
-This document explains how we've structured the repository to add local development tools while keeping them separate from the deployment configuration.
-
-## The Problem
-
-When running a Medusa.js e-commerce site across multiple environments, we faced these challenges:
-
-1. Region ID mismatches between environments causing "Region not found" errors
-2. Database schema differences requiring specific migrations and fixes
-3. Need for easy import of production data into local development
-4. Need to keep local development tools separate from deployment configuration
-
-## Our Solution
-
-We've implemented a two-pronged approach:
-
-### 1. Local Development Tools (Not Deployed to Production)
-
-These files are only used locally and don't affect deployment:
-
-- `setup-local-db.js` - For initial setup and importing production data
-- `restart-dev.sh` - For quick restart and fixing common issues
-- `fix-regions.js` - For fixing region ID mismatches
-- `docker-compose.yml` - For local Docker services
-- Documentation files:
-  - `LOCAL_DEV_README.md`
-  - `REGION_ISSUES.md`
-  - `SOLUTION.md`
-
-All these files contain safety checks to prevent running in a production environment.
-
-### 2. Clean Deployment Files (Used Only for Production)
-
-The core Medusa.js application files remain untouched:
-
-- `backend/` - Standard Medusa backend configuration
-- `storefront/` - Standard Next.js storefront
-
-## Separation of Concerns
-
-To maintain this separation:
-
-1. **Safety Checks**: Development scripts check for `NODE_ENV === "production"` and exit if detected
-2. **No Cross-References**: Deployment code doesn't reference or depend on development scripts
-3. **Clear Documentation**: Each tool's purpose and scope is well-documented
-4. **Railway Deployment**: Follows standard Medusa deployment practices without custom workflows
-
-## Benefits of This Approach
-
-1. **Clean Repository**: Production deployment gets only what it needs
-2. **Easy Onboarding**: New developers can quickly set up with one command
-3. **Simplified Troubleshooting**: Common issues have easy-to-use fix scripts
-4. **Environment Isolation**: Local development tools don't interfere with production
-
-## Conclusion
-
-By clearly separating local development tools from deployment code, we've created a development environment that:
-
-1. Makes it easy to set up and maintain local instances
-2. Solves common region ID and schema issues
-3. Allows for easy production data imports
-4. Doesn't interfere with production deployment
-
-The development tools use Docker Compose to create a consistent environment, while the production deployment relies solely on the standard Medusa deployment process, ensuring that both environments work optimally without interfering with each other.
